+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ### New features
 
-* `data` argument in `init` accepts now `teal_data` and `teal_data_module`.
+* `data` argument in `init` now accepts `teal_data` and `teal_data_module`.
 * Added `landing_popup_module` function which creates a module that will display a popup when the app starts. The popup will block access to the app until it is dismissed.
 * Filter state snapshots can now be uploaded from file. See `?snapshot`.